# [libc++][format][2/7] Optimizes c-string arguments.