The formatter specializations for _CharT* and const _CharT* typically write all elements in a loop. This format's internal functions are optimized for larger writes.

Instead of writing one element at a time, convert the range to a basic_string_view and write that instead.

For C string of 6 characters this is a bit slower, but for 60 characters it's faster. The improvements for back_inserter<std::list<_CharT>> are not as great as the others; it just gets as slow as basic_string_view<_CharT>.

```
Before
---------------------------------------------------------------------------------------------------------------
Benchmark                                                                     Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------
BM_sprintf/C string len = 6                                                4.81 ns         4.80 ns    145890280
BM_format/C string len = 6                                                 52.6 ns         52.4 ns     13252327
BM_format_to_back_inserter<std::string>/C string len = 6                   53.0 ns         52.8 ns     13262680
BM_format_to_back_inserter<std::vector<char>>/C string len = 6             69.7 ns         69.6 ns     10045636
BM_format_to_back_inserter<std::deque<char>>/C string len = 6               148 ns          148 ns      4729368
BM_format_to_back_inserter<std::list<char>>/C string len = 6                127 ns          126 ns      5538441
BM_format_to_iterator/<std::array> C string len = 6                        42.9 ns         42.8 ns     16367158
BM_format_to_iterator/<std::string> C string len = 6                       43.5 ns         43.4 ns     16141644
BM_format_to_iterator/<std::vector> C string len = 6                       42.9 ns         42.8 ns     16366718
BM_format_to_iterator/<std::deque> C string len = 6                        47.8 ns         47.7 ns     14686488
BM_format/string len = 6                                                   55.3 ns         55.2 ns     12696889
BM_format_to_back_inserter<std::string>/string len = 6                     55.4 ns         55.2 ns     12660731
BM_format_to_back_inserter<std::vector<char>>/string len = 6               70.7 ns         70.5 ns      9927313
BM_format_to_back_inserter<std::deque<char>>/string len = 6                 153 ns          153 ns      4573936
BM_format_to_back_inserter<std::list<char>>/string len = 6                  128 ns          128 ns      5486033
BM_format_to_iterator/<std::array> string len = 6                          44.6 ns         44.5 ns     15758122
BM_format_to_iterator/<std::string> string len = 6                         44.7 ns         44.6 ns     15690226
BM_format_to_iterator/<std::vector> string len = 6                         44.3 ns         44.2 ns     15715898
BM_format_to_iterator/<std::deque> string len = 6                          50.3 ns         50.1 ns     13958635
BM_format/string_view len = 6                                              54.2 ns         54.1 ns     12929525
BM_format_to_back_inserter<std::string>/string_view len = 6                54.3 ns         54.1 ns     12929219
BM_format_to_back_inserter<std::vector<char>>/string_view len = 6          70.0 ns         69.8 ns     10022355
BM_format_to_back_inserter<std::deque<char>>/string_view len = 6            153 ns          152 ns      4585749
BM_format_to_back_inserter<std::list<char>>/string_view len = 6             128 ns          128 ns      5489760
BM_format_to_iterator/<std::array> string_view len = 6                     44.2 ns         44.1 ns     15884839
BM_format_to_iterator/<std::string> string_view len = 6                    44.8 ns         44.6 ns     15664278
BM_format_to_iterator/<std::vector> string_view len = 6                    44.7 ns         44.6 ns     15716983
BM_format_to_iterator/<std::deque> string_view len = 6                     50.3 ns         50.2 ns     13936091
BM_sprintf/C string len = 60                                               4.16 ns         4.15 ns    168764227
BM_format/C string len = 60                                                 169 ns          169 ns      4144060
BM_format_to_back_inserter<std::string>/C string len = 60                   167 ns          167 ns      4203915
BM_format_to_back_inserter<std::vector<char>>/C string len = 60             177 ns          176 ns      3965619
BM_format_to_back_inserter<std::deque<char>>/C string len = 60              383 ns          382 ns      1832531
BM_format_to_back_inserter<std::list<char>>/C string len = 60              1270 ns         1267 ns       552686
BM_format_to_iterator/<std::array> C string len = 60                        141 ns          140 ns      4988441
BM_format_to_iterator/<std::string> C string len = 60                       141 ns          141 ns      4956101
BM_format_to_iterator/<std::vector> C string len = 60                       141 ns          141 ns      4963443
BM_format_to_iterator/<std::deque> C string len = 60                        144 ns          143 ns      4893139
BM_format/string len = 60                                                  73.4 ns         73.2 ns      9548455
BM_format_to_back_inserter<std::string>/string len = 60                    73.2 ns         73.0 ns      9524524
BM_format_to_back_inserter<std::vector<char>>/string len = 60              81.6 ns         81.4 ns      8584033
BM_format_to_back_inserter<std::deque<char>>/string len = 60                279 ns          279 ns      2515146
BM_format_to_back_inserter<std::list<char>>/string len = 60                1177 ns         1174 ns       597172
BM_format_to_iterator/<std::array> string len = 60                         44.5 ns         44.4 ns     15753131
BM_format_to_iterator/<std::string> string len = 60                        44.7 ns         44.6 ns     15692630
BM_format_to_iterator/<std::vector> string len = 60                        44.8 ns         44.7 ns     15664689
BM_format_to_iterator/<std::deque> string len = 60                         50.6 ns         50.5 ns     13838617
BM_format/string_view len = 60                                             72.8 ns         72.6 ns      9674007
BM_format_to_back_inserter<std::string>/string_view len = 60               72.7 ns         72.6 ns      9638209
BM_format_to_back_inserter<std::vector<char>>/string_view len = 60         82.6 ns         82.4 ns      8496602
BM_format_to_back_inserter<std::deque<char>>/string_view len = 60           280 ns          280 ns      2508982
BM_format_to_back_inserter<std::list<char>>/string_view len = 60           1176 ns         1173 ns       597714
BM_format_to_iterator/<std::array> string_view len = 60                    44.2 ns         44.1 ns     15896934
BM_format_to_iterator/<std::string> string_view len = 60                   44.7 ns         44.6 ns     15695427
BM_format_to_iterator/<std::vector> string_view len = 60                   44.7 ns         44.6 ns     15680899
BM_format_to_iterator/<std::deque> string_view len = 60                    50.3 ns         50.1 ns     13962755
BM_sprintf/C string len = 6000                                              114 ns          114 ns      6170153
BM_format/C string len = 6000                                             11792 ns        11763 ns        59619
BM_format_to_back_inserter<std::string>/C string len = 6000               11778 ns        11746 ns        59572
BM_format_to_back_inserter<std::vector<char>>/C string len = 6000         11715 ns        11686 ns        60053
BM_format_to_back_inserter<std::deque<char>>/C string len = 6000          25464 ns        25403 ns        27538
BM_format_to_back_inserter<std::list<char>>/C string len = 6000          126044 ns       125726 ns         5572
BM_format_to_iterator/<std::array> C string len = 6000                    10792 ns        10766 ns        64943
BM_format_to_iterator/<std::string> C string len = 6000                   10791 ns        10761 ns        64918
BM_format_to_iterator/<std::vector> C string len = 6000                   10775 ns        10750 ns        65149
BM_format_to_iterator/<std::deque> C string len = 6000                    11242 ns        11215 ns        62542
BM_format/string len = 6000                                                 919 ns          917 ns       763961
BM_format_to_back_inserter<std::string>/string len = 6000                   920 ns          917 ns       763094
BM_format_to_back_inserter<std::vector<char>>/string len = 6000             888 ns          886 ns       793468
BM_format_to_back_inserter<std::deque<char>>/string len = 6000            14999 ns        14963 ns        46758
BM_format_to_back_inserter<std::list<char>>/string len = 6000            114504 ns       114213 ns         6130
BM_format_to_iterator/<std::array> string len = 6000                        120 ns          120 ns      5823426
BM_format_to_iterator/<std::string> string len = 6000                       106 ns          106 ns      6609299
BM_format_to_iterator/<std::vector> string len = 6000                       106 ns          106 ns      6613867
BM_format_to_iterator/<std::deque> string len = 6000                        367 ns          366 ns      1912818
BM_format/string_view len = 6000                                            910 ns          907 ns       771491
BM_format_to_back_inserter<std::string>/string_view len = 6000              911 ns          909 ns       770065
BM_format_to_back_inserter<std::vector<char>>/string_view len = 6000        878 ns          876 ns       794976
BM_format_to_back_inserter<std::deque<char>>/string_view len = 6000       15031 ns        14994 ns        46651
BM_format_to_back_inserter<std::list<char>>/string_view len = 6000       114510 ns       114219 ns         6130
BM_format_to_iterator/<std::array> string_view len = 6000                   121 ns          120 ns      5814249
BM_format_to_iterator/<std::string> string_view len = 6000                  116 ns          115 ns      6072315
BM_format_to_iterator/<std::vector> string_view len = 6000                  115 ns          115 ns      6095762
BM_format_to_iterator/<std::deque> string_view len = 6000                   351 ns          350 ns      2002175

After
---------------------------------------------------------------------------------------------------------------
Benchmark                                                                     Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------
BM_sprintf/C string len = 6                                                4.83 ns         4.83 ns    145502957
BM_format/C string len = 6                                                 55.1 ns         55.1 ns     12687656
BM_format_to_back_inserter<std::string>/C string len = 6                   55.1 ns         55.1 ns     12691642
BM_format_to_back_inserter<std::vector<char>>/C string len = 6             71.2 ns         71.3 ns      9819560
BM_format_to_back_inserter<std::deque<char>>/C string len = 6               154 ns          154 ns      4548709
BM_format_to_back_inserter<std::list<char>>/C string len = 6                131 ns          131 ns      5338318
BM_format_to_iterator/<std::array> C string len = 6                        45.0 ns         45.0 ns     15569429
BM_format_to_iterator/<std::string> C string len = 6                       45.9 ns         45.9 ns     15240594
BM_format_to_iterator/<std::vector> C string len = 6                       44.4 ns         44.4 ns     15768343
BM_format_to_iterator/<std::deque> C string len = 6                        50.1 ns         50.1 ns     13995837
BM_format/string len = 6                                                   54.8 ns         54.8 ns     12738209
BM_format_to_back_inserter<std::string>/string len = 6                     55.4 ns         55.4 ns     12634483
BM_format_to_back_inserter<std::vector<char>>/string len = 6               70.4 ns         70.4 ns      9948288
BM_format_to_back_inserter<std::deque<char>>/string len = 6                 154 ns          154 ns      4531741
BM_format_to_back_inserter<std::list<char>>/string len = 6                  132 ns          132 ns      5289798
BM_format_to_iterator/<std::array> string len = 6                          44.5 ns         44.5 ns     15743437
BM_format_to_iterator/<std::string> string len = 6                         44.9 ns         44.9 ns     15591961
BM_format_to_iterator/<std::vector> string len = 6                         45.0 ns         45.0 ns     15541308
BM_format_to_iterator/<std::deque> string len = 6                          50.5 ns         50.5 ns     13837884
BM_format/string_view len = 6                                              54.5 ns         54.5 ns     12833591
BM_format_to_back_inserter<std::string>/string_view len = 6                54.6 ns         54.6 ns     12818233
BM_format_to_back_inserter<std::vector<char>>/string_view len = 6          69.9 ns         69.9 ns     10018025
BM_format_to_back_inserter<std::deque<char>>/string_view len = 6            154 ns          154 ns      4534215
BM_format_to_back_inserter<std::list<char>>/string_view len = 6             131 ns          131 ns      5337938
BM_format_to_iterator/<std::array> string_view len = 6                     44.2 ns         44.2 ns     15854695
BM_format_to_iterator/<std::string> string_view len = 6                    45.0 ns         45.0 ns     15551054
BM_format_to_iterator/<std::vector> string_view len = 6                    45.0 ns         45.0 ns     15567905
BM_format_to_iterator/<std::deque> string_view len = 6                     50.5 ns         50.5 ns     13858261
BM_sprintf/C string len = 60                                               4.16 ns         4.16 ns    168066307
BM_format/C string len = 60                                                73.8 ns         73.8 ns      9487364
BM_format_to_back_inserter<std::string>/C string len = 60                  73.7 ns         73.7 ns      9492371
BM_format_to_back_inserter<std::vector<char>>/C string len = 60            83.1 ns         83.1 ns      8399017
BM_format_to_back_inserter<std::deque<char>>/C string len = 60              284 ns          284 ns      2466510
BM_format_to_back_inserter<std::list<char>>/C string len = 60              1177 ns         1177 ns       592514
BM_format_to_iterator/<std::array> C string len = 60                       44.9 ns         44.9 ns     15558172
BM_format_to_iterator/<std::string> C string len = 60                      45.8 ns         45.8 ns     15283103
BM_format_to_iterator/<std::vector> C string len = 60                      44.6 ns         44.6 ns     15678967
BM_format_to_iterator/<std::deque> C string len = 60                       50.5 ns         50.5 ns     13839524
BM_format/string len = 60                                                  72.2 ns         72.2 ns      9687634
BM_format_to_back_inserter<std::string>/string len = 60                    72.0 ns         72.1 ns      9695746
BM_format_to_back_inserter<std::vector<char>>/string len = 60              82.3 ns         82.3 ns      8509528
BM_format_to_back_inserter<std::deque<char>>/string len = 60                279 ns          279 ns      2506108
BM_format_to_back_inserter<std::list<char>>/string len = 60                1180 ns         1180 ns       592963
BM_format_to_iterator/<std::array> string len = 60                         44.7 ns         44.7 ns     15657865
BM_format_to_iterator/<std::string> string len = 60                        45.0 ns         45.0 ns     15569049
BM_format_to_iterator/<std::vector> string len = 60                        44.9 ns         44.9 ns     15576187
BM_format_to_iterator/<std::deque> string len = 60                         50.7 ns         50.7 ns     13803188
BM_format/string_view len = 60                                             72.4 ns         72.4 ns      9663061
BM_format_to_back_inserter<std::string>/string_view len = 60               72.8 ns         72.8 ns      9638941
BM_format_to_back_inserter<std::vector<char>>/string_view len = 60         81.9 ns         81.9 ns      8529931
BM_format_to_back_inserter<std::deque<char>>/string_view len = 60           283 ns          283 ns      2471522
BM_format_to_back_inserter<std::list<char>>/string_view len = 60           1177 ns         1177 ns       594752
BM_format_to_iterator/<std::array> string_view len = 60                    44.2 ns         44.2 ns     15841413
BM_format_to_iterator/<std::string> string_view len = 60                   45.1 ns         45.1 ns     15546279
BM_format_to_iterator/<std::vector> string_view len = 60                   45.0 ns         45.0 ns     15555756
BM_format_to_iterator/<std::deque> string_view len = 60                    50.1 ns         50.1 ns     10000000
BM_sprintf/C string len = 6000                                              344 ns          344 ns      2037753
BM_format/C string len = 6000                                               980 ns          980 ns       713341
BM_format_to_back_inserter<std::string>/C string len = 6000                 979 ns          979 ns       709571
BM_format_to_back_inserter<std::vector<char>>/C string len = 6000           944 ns          944 ns       744518
BM_format_to_back_inserter<std::deque<char>>/C string len = 6000          14916 ns        14915 ns        46875
BM_format_to_back_inserter<std::list<char>>/C string len = 6000          114686 ns       114688 ns         6110
BM_format_to_iterator/<std::array> C string len = 6000                      159 ns          159 ns      4394506
BM_format_to_iterator/<std::string> C string len = 6000                     159 ns          159 ns      4384577
BM_format_to_iterator/<std::vector> C string len = 6000                     159 ns          159 ns      4393225
BM_format_to_iterator/<std::deque> C string len = 6000                      437 ns          437 ns      1601253
BM_format/string len = 6000                                                 929 ns          929 ns       751889
BM_format_to_back_inserter<std::string>/string len = 6000                   933 ns          933 ns       752042
BM_format_to_back_inserter<std::vector<char>>/string len = 6000             892 ns          892 ns       785064
BM_format_to_back_inserter<std::deque<char>>/string len = 6000            14840 ns        14839 ns        47177
BM_format_to_back_inserter<std::list<char>>/string len = 6000            114940 ns       114942 ns         6095
BM_format_to_iterator/<std::array> string len = 6000                        144 ns          144 ns      4859849
BM_format_to_iterator/<std::string> string len = 6000                       106 ns          106 ns      6565564
BM_format_to_iterator/<std::vector> string len = 6000                       107 ns          107 ns      6584152
BM_format_to_iterator/<std::deque> string len = 6000                        368 ns          368 ns      1904727
BM_format/string_view len = 6000                                            919 ns          919 ns       760401
BM_format_to_back_inserter<std::string>/string_view len = 6000              917 ns          917 ns       760934
BM_format_to_back_inserter<std::vector<char>>/string_view len = 6000        880 ns          880 ns       796244
BM_format_to_back_inserter<std::deque<char>>/string_view len = 6000       14854 ns        14853 ns        47101
BM_format_to_back_inserter<std::list<char>>/string_view len = 6000       114963 ns       114967 ns         6086
BM_format_to_iterator/<std::array> string_view len = 6000                   117 ns          117 ns      5974080
BM_format_to_iterator/<std::string> string_view len = 6000                  116 ns          116 ns      6016080
…