@philnik777 2024-09-17
It's also worth noting that compilers can constant-fold `strlen`, so in cases where the string is known this doesn't even result in multiple iterations.